backend/src/requirements.txt
asgiref==3.7.2
Django==5.0.1
djangorestframework==3.14.0
gunicorn==21.2.0
packaging==23.2
pytz==2023.3.post1
sqlparse==0.4.4
requests==2.31.0

# PostgresSQL
psycopg2-binary>=2.9

# CORS Headers on REST
django-cors-headers==4.3.1

# Scrapping
beautifulsoup4==4.12.3
